Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@5180 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ checksum/src/nu/mine/mosher/checksum/StreamChecksum.java
@@ -33,7 +33,7 @@ public class ChecksumStream
         byte[] rs = new byte[cWindow];
         RollingChecksum rollCheck = new RollingChecksum();
 
-        int w = 0;
+        long w = 0;
         while (is.available() > 0)
         {
             if (is.available() < cWindow)
